valkey: reuse Redis client, override only time series
- Embed Redis struct in RedisValkey, inherit all methods unchanged
- Override AddTimeSeriesPoints/GetTimeSeriesPoints with sorted-set + hash
- Register valkey://, valkeys://, valkey+cluster://, valkeys+cluster:// schemes
- Remove standalone valkey.go (~839 lines) and valkey-go dependency
- Add TestValkeyViaRedis and TestRedisSortedSetTS test suites
